requirements.txt
pip~=23.2.1
wheel~=0.41.2
setuptools~=68.2.0
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3
python-dotenv==1.0.0